requirements.txt
Flask
Flask-PyMongo
requests
python-dotenv
schedule
